% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualise.r
\name{align}
\alias{align}
\title{Align line profiling data with source code}
\usage{
align(x, digits = 3)
}
\arguments{
\item{x}{a line profiling object}

\item{digits}{number of digits to use for performance summaries}
}
\description{
Align line profiling data with source code
}

