% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.lineprof}
\alias{summary.lineprof}
\title{Summarize call stack by function}
\usage{
\method{summary}{lineprof}(x, depth = 1)
}
\arguments{
\item{x}{a line profiling object}

\item{depth}{depth of calls}
}
\description{
Aggregates the calling stack bottom-up by function name.
}
\examples{
\dontrun{
library(jsonlite)
library(ggplot2)
l1 <- lineprof(json <- toJSON(diamonds))
#the usual top-down aggregation
print(l1)

#bottom-up aggregation
summary(l1, depth = 3)
}
}

